reorganize kalosm-language
