+++ interfaces/kalosm-language/src/context/document.rs
@@ -72,3 +72,25 @@ impl IntoDocument for Url {
         super::page::get_article(self).await
     }
 }
+
+/// A document that can be added to a search index.
+#[async_trait::async_trait]
+pub trait IntoDocuments {
+    /// Convert the document into a [`Document`]
+    async fn into_documents(self) -> anyhow::Result<Vec<Document>>;
+}
+
+#[async_trait::async_trait]
+impl<T: IntoDocument + Send + Sync, I> IntoDocuments for I
+where
+    I: IntoIterator<Item = T> + Send + Sync,
+    <I as IntoIterator>::IntoIter: Send + Sync,
+{
+    async fn into_documents(self) -> anyhow::Result<Vec<Document>> {
+        let mut documents = Vec::new();
+        for document in self {
+            documents.push(document.into_document().await?);
+        }
+        Ok(documents)
+    }
+}
