+++ interfaces/kalosm-language/src/context/io/mod.rs
@@ -1,6 +1,6 @@
 use crate::context::document::Document;
 use crate::context::document::IntoDocument;
-use crate::index::IntoDocuments;
+use crate::context::document::IntoDocuments;
 use std::path::PathBuf;
 use tokio::task::JoinSet;
 mod docx;
