+++ interfaces/kalosm-streams/src/timed_stream.rs
@@ -1,3 +1,5 @@
+//! Streams for time series data.
+
 use std::collections::VecDeque;
 
 use futures_util::Stream;
