+++ interfaces/kalosm/examples/chat-mistral-2.rs
@@ -1,4 +1,4 @@
-use kalosm::{language::*, *};
+use kalosm::language::*;
 
 #[tokio::main]
 async fn main() {
