+++ interfaces/kalosm/examples/chat.rs
@@ -1,4 +1,4 @@
-use kalosm::{language::*, *};
+use kalosm::language::*;
 
 #[tokio::main]
 async fn main() {
