+++ interfaces/kalosm/examples/live_qa.rs
@@ -1,6 +1,4 @@
-use futures_util::StreamExt;
-use kalosm::*;
-use kalosm_language::*;
+use kalosm::language::*;
 use kalosm_sound::*;
 use std::sync::Arc;
 use tokio::{
