+++ interfaces/kalosm/examples/phi.rs
@@ -1,4 +1,4 @@
-use kalosm_language::*;
+use kalosm::language::*;
 
 #[tokio::main]
 async fn main() {
