+++ interfaces/kalosm/examples/remote.rs
@@ -1,10 +1,6 @@
 // You must set the environment variable OPENAI_API_KEY (https://platform.openai.com/account/api-keys) to run this example.
 
-use std::io::Write;
-
-use futures_util::stream::StreamExt;
-use kalosm_language::*;
-use kalosm_streams::TextStream;
+use kalosm::language::*;
 
 #[tokio::main]
 async fn main() {
@@ -15,10 +11,5 @@ async fn main() {
     print!("{}", prompt);
 
     let stream = llm.stream_text(prompt).with_max_length(300).await.unwrap();
-
-    let mut sentences = stream.words();
-    while let Some(text) = sentences.next().await {
-        print!("{}", text);
-        std::io::stdout().flush().unwrap();
-    }
+    stream.to_std_out().await.unwrap();
 }
