+++ interfaces/kalosm/examples/remote_embed.rs
@@ -2,8 +2,7 @@
 
 use std::io::Write;
 
-use futures_util::stream::StreamExt;
-use kalosm_language::*;
+use kalosm::language::*;
 
 #[tokio::main]
 async fn main() {
