+++ models/rmistral/src/language_model.rs
@@ -4,7 +4,7 @@ pub use crate::Mistral;
 use crate::MistralModel;
 use crate::{InferenceSettings, Task};
 use kalosm_language_model::{CreateModel, GenerationParameters, Model, VectorSpace};
-use kalosm_streams::ChannelTextStream;
+use kalosm_streams::text_stream::ChannelTextStream;
 
 #[async_trait::async_trait]
 impl CreateModel for Mistral {
